refactoring: 타입 폴더 구조 수정
