+++ src/entities/auction/index.ts
@@ -3,3 +3,4 @@ export type * from './item';
 export type * from './list';
 export type * from './userParticipated';
 export type * from './userRegistered';
+export type * from './myAuctionData';
