+++ src/features/address/api/addAddress.ts
@@ -1,4 +1,4 @@
-import { IAddressBase } from "@/@types/Address";
+import { IAddressBase } from "@/entities/address/address";
 import { API_END_POINT, httpClient } from "@/shared";
 
 export const addAddress = async (data: IAddressBase) => {
