+++ src/features/user/api/getAuctionOngoingRegister.ts
@@ -1,6 +1,6 @@
 import { API_END_POINT, httpClient } from '@/shared';
 
-import { GetAuctionProps } from '@/@types/user';
+import { GetAuctionProps } from '@/entities/user/user';
 import type { IAuctionOngoingRegisteredList } from '@/entities';
 
 export const getAuctionOngoingRegister = async ({ pageNumber, pageSize, sortType = 'newest' }: GetAuctionProps): Promise<IAuctionOngoingRegisteredList> => {
