+++ src/features/user/api/getMyLostAuction.ts
@@ -1,6 +1,6 @@
 import { API_END_POINT, httpClient } from '@/shared';
 
-import { GetAuctionProps } from '@/@types/user';
+import { GetAuctionProps } from '@/entities/user/user';
 import type { IUserAuctionLostList } from '@/entities';
 
 export const getMyLostAuction = async ({ pageNumber, pageSize }: GetAuctionProps): Promise<IUserAuctionLostList> => {
